\name{intasymmRcppArma-package}
\alias{intasymmRcppArma-package}
\alias{intasymmRcppArma}
\docType{package}
\title{
\packageTitle{intasymmRcppArma}
}
\description{
\packageDescription{intasymmRcppArma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{intasymmRcppArma}
\packageIndices{intasymmRcppArma}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{intasymmRcppArma}

Maintainer: \packageMaintainer{intasymmRcppArma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
